+++ cpp/main.cpp
@@ -1,168 +1,12 @@
-#include "yolov5.h"
+#include "yolo.h"
 
 
 int main(int argc, char* argv[])
 {
-	std::cout << "Libtorch CPU FP32" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::Libtorch);
-		yolov5->init("yolov5n_cpu_fp32.torchscript", CPU, FP32);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "Libtorch GPU FP32" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::Libtorch);
-		yolov5->init("yolov5n_gpu_fp32.torchscript", GPU, FP32);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "Libtorch GPU FP16" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::Libtorch);
-		yolov5->init("yolov5n_gpu_fp16.torchscript", GPU, FP16);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "ONNXRuntime CPU FP32" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::ONNXRuntime);
-		yolov5->init("yolov5n_fp32.onnx", CPU, FP32);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "ONNXRuntime GPU FP32" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::ONNXRuntime);
-		yolov5->init("yolov5n_fp32.onnx", GPU, FP32);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "ONNXRuntime CPU FP16" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::ONNXRuntime);
-		yolov5->init("yolov5n_fp16.onnx", CPU, FP16);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "ONNXRuntime GPU FP16" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::ONNXRuntime);
-		yolov5->init("yolov5n_fp16.onnx", GPU, FP16);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "ONNXRuntime CPU INT8" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::ONNXRuntime);
-		yolov5->init("yolov5n_int8.onnx", CPU, INT8);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "ONNXRuntime GPU INT8" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::ONNXRuntime);
-		yolov5->init("yolov5n_int8.onnx", GPU, INT8);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "OpenCV CPU FP32" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::OpenCV);
-		yolov5->init("yolov5n_fp32.onnx", CPU, FP32);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "OpenCV GPU FP32" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::OpenCV);
-		yolov5->init("yolov5n_fp32.onnx", GPU, FP32);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "OpenVINO CPU FP32" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::OpenVINO);
-		yolov5->init("yolov5n_fp32.xml", CPU, FP32);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "OpenVINO GPU FP32" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::OpenVINO);
-		yolov5->init("yolov5n_fp32.xml", GPU, FP32);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "OpenVINO CPU FP16" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::OpenVINO);
-		yolov5->init("yolov5n_fp16.xml", CPU, FP16);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "OpenVINO GPU FP16" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::OpenVINO);
-		yolov5->init("yolov5n_fp16.xml", GPU, FP16);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "OpenVINO CPU INT8" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::OpenVINO);
-		yolov5->init("yolov5n_int8.xml", CPU, INT8);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "OpenVINO GPU INT8" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::OpenVINO);
-		yolov5->init("yolov5n_int8.xml", GPU, INT8);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "TensorRT GPU FP32" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::TensorRT);
-		yolov5->init("yolov5n_fp32.engine", GPU, FP32);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "TensorRT GPU FP16" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::TensorRT);
-		yolov5->init("yolov5n_fp16.engine", GPU, FP16);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
-
-	std::cout << "TensorRT GPU INT8" << std::endl;
-	{
-		std::unique_ptr<YOLOv5> yolov5 = AlgoFactory::instance().create(Algo_Type::TensorRT);
-		yolov5->init("yolov5n_int8.engine", GPU, INT8);
-		yolov5->infer("test.mp4");
-		yolov5->release();
-	}
+	std::unique_ptr<YOLO> yolo = BackendFactory::instance().create(Backend_Type(atoi(argv[1])));
+	yolo->init(Algo_Type(atoi(argv[2])), Device_Type(atoi(argv[3])), Model_Type(atoi(argv[4])), argv[5]);
+	yolo->infer("test.mp4", argv);
+	yolo->release();
 
 	return 0;
 }
-
