--- cpp/yolov5_libtorch.h
@@ -1,28 +1,0 @@
-#pragma once
-
-
-#include "yolov5.h"
-#include <torch/script.h>
-#include <torch/torch.h>
-
-
-class YOLOv5_Libtorch : public YOLOv5
-{	
-public:
-	void init(const std::string model_path, const Device_Type device_type, const Model_Type model_type);
-
-private:
-	void pre_process();
-		
-	void process();
-
-	torch::DeviceType m_device;
-
-	Model_Type m_model;
-
-	torch::jit::script::Module module;
-
-	std::vector<torch::jit::IValue> m_inputs;
-
-	torch::jit::IValue m_outputs;
-};
